Storybook: Split Examples and API for Area component as an example. Show code for API and show story source for Examples.
