+++ storybook/main.js
@@ -11,9 +11,9 @@ module.exports = {
     '@storybook/addon-links',
     '@storybook/addon-essentials',
     '@storybook/addon-interactions',
-    '@storybook/addon-docs',
     '@storybook/addon-a11y',
-    '@storybook/addon-mdx-gfm',
+    '@storybook/addon-storysource',
+    '@storybook/addon-docs',
   ],
   framework: {
     name: '@storybook/react-webpack5',
@@ -23,6 +23,6 @@ module.exports = {
     interactionsDebugger: true,
   },
   docs: {
-    autodocs: true,
+    autodocs: false,
   },
 };
