+++ storybook/stories/API/cartesian/Area.mdx
@@ -1,7 +1,13 @@
-import { ArgTypes, Meta, Stories } from '@storybook/blocks';
+import { ArgTypes, Meta, Story, Canvas, Controls } from '@storybook/blocks';
 import * as ComponentStories from './Area.stories';
 
 # Area
+
+<Meta of={ComponentStories} />
+
+<Canvas>
+  <Story name="API" of={ComponentStories} />
+</Canvas>
 
 ## Parent Component
 The Area can be used within a `<ComposedChart/>` or a `<AreaChart/>`.
@@ -10,11 +16,14 @@ The Area can be used within a `<ComposedChart/>` or a `<AreaChart/>`.
 The Area can be used with the following child components: `<LabelList/>`
 
 ## Properties
+The following table shows a complete list of all properties of the Area.
 Properties in the groups Other and Internal are not recommended to be used.
 
-<Meta of={ComponentStories} />;
-<ArgTypes of={ComponentStories} sort={'requiredFirst'} />;
-<Stories />;
+<Controls of={ComponentStories.API} />
+
+
+
+
