+++ storybook/stories/Examples/cartesian/Area/AreaWithFillPattern.stories.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import { ComposedChart, Area, ResponsiveContainer } from '../../../../../src';
+import { coordinateWithValueData } from '../../../data';
+
+export default {
+  title: 'Examples/cartesian/Area/With Fill Pattern',
+};
+
+const [surfaceWidth, surfaceHeight] = [600, 300];
+
+export const FillPattern = {
+  render: () => {
+    return (
+      <ResponsiveContainer width="100%" height={surfaceHeight}>
+        <ComposedChart
+          width={surfaceWidth}
+          height={surfaceHeight}
+          margin={{
+            top: 20,
+            right: 20,
+            bottom: 20,
+            left: 20,
+          }}
+          data={coordinateWithValueData}
+        >
+          <defs>
+            <pattern id="left" width="12" height="4" patternUnits="userSpaceOnUse">
+              <rect width="4" height="4" fill="#8884d8" />
+            </pattern>
+            <pattern id="right" width="8" height="4" patternUnits="userSpaceOnUse" patternTransform="rotate(45)">
+              <rect width="4" height="4" fill="#82ca9d" />
+            </pattern>
+          </defs>
+          <Area type="monotone" dataKey="x" stroke="#8884d8" fillOpacity={1} fill="url(#left)" />
+          <Area type="monotone" dataKey="y" stroke="#82ca9d" fillOpacity={1} fill="url(#right)" />
+        </ComposedChart>
+      </ResponsiveContainer>
+    );
+  },
+};
